+++ sites/all/modules/custom/mck_venn_diagram/js/mck_venn_diagram_magnific_popup.js
@@ -27,6 +27,43 @@
               if (this.arrowLeft && this.arrowRight) {
                 this.contentContainer.append(this.arrowLeft.add(this.arrowRight));
               }
+            },
+            open: function() {
+              var mfp = $.magnificPopup.instance;
+              var proto = $.magnificPopup.proto;
+
+              function updateArrow () {
+                if(mfp.index < mfp.items.length - 1) {
+                  $('.mfp-arrow-right').show();
+                }
+                else {
+                  $('.mfp-arrow-right').hide();
+                }
+
+                if(mfp.index > 0) {
+                  $('.mfp-arrow-left').show();
+                }
+                else {
+                  $('.mfp-arrow-left').hide();
+                }
+              }
+
+              // extend function that moves to next item
+              mfp.next = function() {
+                if(mfp.index < mfp.items.length - 1) {
+                  proto.next.call(mfp);
+                }
+                updateArrow();
+              };
+              // same with prev method
+              mfp.prev = function() {
+                if(mfp.index > 0) {
+                  proto.prev.call(mfp);
+                }
+                updateArrow();
+              };
+
+              updateArrow();
             }
           },
           gallery: {